fixed render function
remove function for adding genres to localStorage to another function
